requirements.txt
streamlit
pillow
numpy
matplotlib
opencv-python
reportlab
tensorflow==2.18.0
pandas
gdown
